configurable script to transfer usdc to your wallet + readme instructions on running
