+++ scripts/run-testnet.sh
@@ -6,6 +6,7 @@ if [ -f .env.local ]; then
 fi
 
 yarn run ganache \
+  --deterministic
 	--fork.url=https://eth-mainnet.alchemyapi.io/v2/$ALCHEMY_API_KEY \
 	--miner.defaultGasPrice 30000000000 \
 	--chain.vmErrorsOnRPCResponse=true \
@@ -15,5 +16,3 @@ yarn run ganache \
 # TODO: transfer USDC from whale to 
 # USDC whale 0xCFFAd3200574698b78f32232aa9D63eABD290703
 # DAI 0x16b34ce9a6a6f7fc2dd25ba59bf7308e7b38e186
-
-curl -H 'Content-Type: application/json' --data '{"jsonrpc":"2.0", "id": 1, "method": "eth_getBalance", "params": ["<Account Address Here>"] }' http://localhost:8545
